Adjust spacing. Rename CSS modules.
